SP, HP 작업 중
SP, HP 증가, 감소 기능은 추가 완료.
UI 작업 해야 함

+++ RandomDice/Assets/Scripts/ScriptableObjects/Monster/Type/MonsterType.cs
@@ -11,5 +11,5 @@ public class MonsterType : ScriptableObject
 
     public float moveSpeedMult = 1f;
 
-    public float attackPowerMult = 1f;
+    public int attackPowerMult = 1;
 }
